+++ src/Chirp.Razor/Chirp.Razor.csproj
@@ -6,4 +6,8 @@
     <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
 
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Data.Sqlite" Version="7.0.11" />
+  </ItemGroup>
+
 </Project>
