+++ src/Chirp.Razor/Chirp.Razor.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.002.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Chirp.Razor", "Chirp.Razor.csproj", "{313A87D4-AC33-4295-A803-1C686798102B}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{313A87D4-AC33-4295-A803-1C686798102B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{313A87D4-AC33-4295-A803-1C686798102B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{313A87D4-AC33-4295-A803-1C686798102B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{313A87D4-AC33-4295-A803-1C686798102B}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {C59A4D08-1E61-4C0C-B44C-CFDD901292FF}
+	EndGlobalSection
+EndGlobal
