+++ src/Chirp.Razor/wwwroot/favicon/favicon.ico
